packages/workflows/pyproject.toml
[project]
name = "autoswarm-workflows"
version = "0.1.0"
description = "Dynamic workflow engine with YAML-to-LangGraph compilation"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.10.5",
  "pyyaml>=6.0",
  "langgraph>=0.2.60",
  "langchain-core>=0.3.28",
  "autoswarm-inference",
  "autoswarm-permissions",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autoswarm_workflows"]
